Chapter7/06-ChatGPT-Google-Map-2.0/requirements.txt
line-bot-sdk==2.4.1
Flask==2.2.3
python-dotenv==0.21.1
openai==0.27.0